naicaweb/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rosetta==0.9.8
idna==3.4
phonenumbers==8.13.1
Pillow==9.3.0
polib==1.1.1
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
